services/auth-service/requirements.txt
# Auth Service Dependencies - Clean Microservice

# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
cryptography>=41.0.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0

# HTTP Client
httpx==0.25.2

# Utilities
python-dotenv==1.0.0

# Additional dependencies for auth service
python-multipart==0.0.6